Evaluate position also based on board control and game result
Each square of the chess board has a value associated with it, with
central squares receiving a greater value. Award evaluation points to
white and black for every legal non-pawn move to a square. Award double
points for squares diagonal to pawns (the squares the pawn could capture
on).

Set evaluation to +1000 for white checkmate and -1000 for black
checkmate. Set evaluation to 0 for any drawn position.